ai-service/requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0

# Hugging Face Tasks
sentence-transformers>=2.2.0
pytesseract>=0.3.10
Pillow>=9.5.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Machine Learning
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
xgboost>=1.7.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
